+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Support for referencing the outputs of other Pulumi stacks has been added to the Pulumi Python libraries via the
   `StackReference` type.
 - Add CI system detection for Bitbucket Pipelines.
+- Pulumi now tolerates changes in default providers in certain cases, which fixes an issue where users would see
+  unexpected replaces when upgrading a Pulumi package.
 - Add support for renaming resources via the `aliases` resource option.  Adding aliases allows new resources to match
   resources from previous deployments which used different names, maintaining the identity of the resource and avoiding
   replacements or re-creation of the resource.